+++ include/image.h
@@ -15,12 +15,12 @@ private:
 public:
     Image();
     void clear();
-    unsigned int get_width();
-    unsigned int get_height();
-    unsigned int load(const char* file_name);
-    unsigned int save(const char* file_name);
-    Color get_pixel(int x, int y);
-    void set_pixel(Color color, int x, int y);
+    size_t get_width();
+    size_t get_height();
+    size_t load(const char* file_name);
+    size_t save(const char* file_name);
+    Color get_pixel(unsigned int x, unsigned int y);
+    void set_pixel(Color color, unsigned int x, unsigned int y);
     std::vector<std::vector<int>> gradient_map();
     void create_from_threshold_matrix(std::vector<std::vector<int>> threshold_matrix);
 };
